+++ tests/QueryType/Update/Query/Document/DocumentTest.php
@@ -26,7 +26,7 @@ class DocumentTest extends TestCase
 
     public function testConstructorWithFieldsAndBoostsAndModifiers()
     {
-        $fields = ['id' => 1, 'name' => 'testname'];
+        $fields = ['id' => 1, 'name' => 'testname', 'categories' => [4, 5]];
         $boosts = ['name' => 2.7];
         $modifiers = ['name' => Document::MODIFIER_SET];
         $doc = new Document($fields, $boosts, $modifiers);
@@ -163,6 +163,54 @@ class DocumentTest extends TestCase
         );
     }
 
+    public function testSetFields()
+    {
+        $this->doc->addField('foo', 'bar');
+        $this->doc->setFieldBoost('name', 2.7);
+        $this->doc->setFieldModifier('categories', Document::MODIFIER_ADD);
+
+        $fields = ['id' => 1, 'name' => 'testname', 'categories' => [4, 5]];
+        $this->doc->setFields($fields);
+
+        $this->assertSame(
+            $fields,
+            $this->doc->getFields()
+        );
+
+        $this->assertNull(
+            $this->doc->getFieldBoost('name')
+        );
+
+        $this->assertNull(
+            $this->doc->getFieldModifier('categories')
+        );
+    }
+
+    public function testSetFieldsWithBoostsAndModifiers()
+    {
+        $fields = ['id' => 1, 'name' => 'testname', 'categories' => [4, 5]];
+        $boosts = ['name' => 2.7];
+        $modifiers = ['categories' => Document::MODIFIER_SET];
+        
+        $this->doc->setKey('id');
+        $this->doc->setFields($fields, $boosts, $modifiers);
+
+        $this->assertSame(
+            $fields,
+            $this->doc->getFields()
+        );
+
+        $this->assertSame(
+            2.7,
+            $this->doc->getFieldBoost('name')
+        );
+
+        $this->assertSame(
+            Document::MODIFIER_SET,
+            $this->doc->getFieldModifier('categories')
+        );
+    }
+
     public function testRemoveField()
     {
         $this->doc->removeField('name');
@@ -207,6 +255,16 @@ class DocumentTest extends TestCase
     public function testRemoveFieldBoostRemoval()
     {
         $this->doc->setFieldBoost('name', 3.2);
+        $this->doc->removeField('name');
+
+        $this->assertNull(
+            $this->doc->getFieldBoost('name')
+        );
+    }
+
+    public function testRemoveFieldModifierRemoval()
+    {
+        $this->doc->setFieldModifier('name', Document::MODIFIER_ADD);
         $this->doc->removeField('name');
 
         $this->assertNull(
@@ -426,11 +484,46 @@ class DocumentTest extends TestCase
         );
     }
 
+    public function testEscapeByDefaultConstructor()
+    {
+        $fields = [
+            'foo' => 'bar'.chr(15),
+            'foos' => ['bar'.chr(15), 'bar'.chr(15).chr(8)],
+        ];
+        $doc = new Document($fields);
+
+        $this->assertSame('bar ', $doc->foo);
+        $this->assertSame(['bar ', 'bar  '], $doc->foos);
+    }
+
+    public function testEscapeByDefaultSetByProperty()
+    {
+        $this->doc->foo = 'bar'.chr(15);
+        $this->doc->foos = ['bar'.chr(15), 'bar'.chr(15).chr(8)];
+
+        $this->assertSame('bar ', $this->doc->foo);
+        $this->assertSame(['bar ', 'bar  '], $this->doc->foos);
+    }
+
     public function testEscapeByDefaultSetField()
     {
         $this->doc->setField('foo', 'bar'.chr(15));
+        $this->doc->setField('foos', ['bar'.chr(15), 'bar'.chr(15).chr(8)]);
 
         $this->assertSame('bar ', $this->doc->foo);
+        $this->assertSame(['bar ', 'bar  '], $this->doc->foos);
+    }
+
+    public function testEscapeByDefaultSetFields()
+    {
+        $fields = [
+            'foo' => 'bar'.chr(15),
+            'foos' => ['bar'.chr(15), 'bar'.chr(15).chr(8)],
+        ];
+        $this->doc->setFields($fields);
+
+        $this->assertSame('bar ', $this->doc->foo);
+        $this->assertSame(['bar ', 'bar  '], $this->doc->foos);
     }
 
     public function testEscapeByDefaultAddField()
@@ -441,12 +534,37 @@ class DocumentTest extends TestCase
         $this->assertSame(['bar ', 'bar  '], $this->doc->foo);
     }
 
+    public function testNoEscapeSetByProperty()
+    {
+        $this->doc->setFilterControlCharacters(false);
+        $this->doc->foo = $value = 'bar'.chr(15);
+        $this->doc->foos = $multivalue = ['bar'.chr(15), 'bar'.chr(15).chr(8)];
+
+        $this->assertSame($value, $this->doc->foo);
+        $this->assertSame($multivalue, $this->doc->foos);
+    }
+
     public function testNoEscapeSetField()
     {
         $this->doc->setFilterControlCharacters(false);
         $this->doc->setField('foo', $value = 'bar'.chr(15));
+        $this->doc->setField('foos', $multivalue = ['bar'.chr(15), 'bar'.chr(15).chr(8)]);
 
         $this->assertSame($value, $this->doc->foo);
+        $this->assertSame($multivalue, $this->doc->foos);
+    }
+
+    public function testNoEscapeSetFields()
+    {
+        $fields = [
+            'foo' => $value = 'bar'.chr(15),
+            'foos' => $multivalue = ['bar'.chr(15), 'bar'.chr(15).chr(8)],
+        ];
+        $this->doc->setFilterControlCharacters(false);
+        $this->doc->setFields($fields);
+
+        $this->assertSame($value, $this->doc->foo);
+        $this->assertSame($multivalue, $this->doc->foos);
     }
 
     public function testNoEscapeAddField()