+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [unreleased]
+### Added
+- QueryType\Update\Query\Document::setFields() to set all fields on a Document
+
+### Fixed
+- Always respect automatic filtering of control characters in field values in QueryType\Update\Query\Document
+
+
 ## [6.1.5]
 ### Added
 - Component\Result\Stats\Result::getDistinctValues()